start using dataframe to processing error curve
