+++ src/EMIRT.jl
@@ -2,17 +2,17 @@ VERSION >=v"0.4.0-dev+6521" && __precompile__()
 
 module EMIRT
 
-include("domains.jl")
-include("affinity.jl")
-#include("boundary.jl")
-include("emshow.jl")
-include("emio.jl")
-include("label.jl")
-include("evaluate.jl")
-include("emparser.jl")
-include("filesystem.jl")
-include("utils.jl")
-include("types.jl")
-include("aws.jl")
+include("core/domains.jl")
+include("core/affinity.jl")
+include("core/boundary.jl")
+#include("emshow.jl")
+include("core/emio.jl")
+include("core/label.jl")
+include("core/evaluate.jl")
+include("core/emparser.jl")
+include("core/filesystem.jl")
+include("core/utils.jl")
+include("core/types.jl")
+#include("aws.jl")
 
 end
