--- src/emparallel.jl
@@ -1,11 +1,0 @@
-# This function retuns the (irange,jrange) indexes assigned to this worker
-@everywhere function myrange(q::SharedArray)
-    idx = indexpids(q)
-    if idx == 0
-        # This worker is not assigned a piece
-        return 1:0, 1:0
-    end
-    nchunks = length(procs(q))
-    splits = [round(Int, s) for s in linspace(0,size(q,2),nchunks+1)]
-    1:size(q,1), splits[idx]+1:splits[idx+1]
-end
