--- src/filesystem.jl
@@ -1,10 +1,0 @@
-export cleardir
-
-"""
-clear a floder
-"""
-function cleardir(dir::AbstractString)
-    for fname in readdir(dir)
-        rm(joinpath(dir, fname), recursive=true)
-    end
-end
